requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
python-decouple==3.8

# Essential AI/ML Dependencies
opencv-python-headless==4.8.1.78
numpy==1.24.3

# NudeNet for real nudity detection
nudenet==3.4.2 